minor improvements

+++ install.sh
@@ -8,7 +8,16 @@ LIB_INSTALL_PATH=/usr/local/lib
 
 WORKSPACE_DIR=$HOME/.$PROG_NAME
 
-python3 -m venv env
+if [[ ! -d "./env" ]]; then
+    echo "Creating a Python virtual environment."
+    python3 -m venv ./env
+else
+    echo "An existing Python virtual environment was detected."
+fi
+
+if [ -e "./dist/$PROG_NAME" ]; then
+    rm ./dist/$PROG_NAME
+fi
 
 source ./env/bin/activate
 
@@ -23,12 +32,27 @@ if [ ! -e "./dist/$PROG_NAME" ]; then
     exit 1
 fi
 
+if [ -e "$PROG_INSTALL_PATH/$PROG_NAME" ]; then
+    echo "Removing a previously installed executable file."
+    sudo rm $PROG_INSTALL_PATH/$PROG_NAME
+fi
+
+echo "Installing the compiled executable file."
 sudo cp ./dist/$PROG_NAME $PROG_INSTALL_PATH
 
-sudo cp ./lib/$LIB_NAME $LIB_INSTALL_PATH
+if [ ! -e "$LIB_INSTALL_PATH/$LIB_NAME" ]; then
+    echo "Installing the RKLLM library."
+    sudo cp ./lib/$LIB_NAME $LIB_INSTALL_PATH
+else
+    echo "The RKLLM library is already installed."
+fi
 
-if [[ ! -d "${WORKSPACE_DIR}" ]]; then
-    mkdir -p ${WORKSPACE_DIR}
+if [[ ! -d "$WORKSPACE_DIR" ]]; then
+    echo "Creating a working directory."
+    mkdir $WORKSPACE_DIR
+    mkdir $WORKSPACE_DIR/models
+else
+    echo "An existing working directory was found."
 fi
 
 echo "Install success."
